+++ src/moveDefinition.ts
@@ -82,8 +82,8 @@ export async function moveDefinitionToMatchingSourceFile(
     if (!declaration && SourceFile.isHeader(definition.uri)) {
         const newDeclaration = definition.newFunctionDeclaration();
         const replaceRange = cfg.alwaysMoveComments(definition.uri)
-                ? definition.rangeWithLeadingComment()
-                : definition.fullRange();
+                ? definition.rangeWithComments
+                : definition.fullRange;
         workspaceEdit.replace(definition.uri, replaceRange, newDeclaration);
     } else {
         const deletionRange = getDeletionRange(definition);
@@ -118,7 +118,7 @@ export async function moveDefinitionIntoOrOutOfClass(
         }
         definition = symbol;
 
-        if (definition.parent?.isClassOrStruct()) {
+        if (definition.parent?.isClassType()) {
             classDoc = sourceDoc;
         } else {
             const declarationLocation = await definition.findDeclaration();
@@ -129,12 +129,12 @@ export async function moveDefinitionIntoOrOutOfClass(
                         ? sourceDoc
                         : await SourceDocument.open(declarationLocation.uri);
                 declaration = await classDoc.getSymbol(declarationLocation.range.start);
-                if (!declaration?.parent?.isClassOrStruct()) {
+                if (!declaration?.parent?.isClassType()) {
                     declaration = undefined;
                 }
             }
 
-            if (declaration?.parent?.isClassOrStruct() === false || !classDoc) {
+            if (declaration?.parent?.isClassType() === false || !classDoc) {
                 const parentClass = await definition.getParentClass();
                 if (parentClass) {
                     classDoc = parentClass.document;
@@ -146,7 +146,7 @@ export async function moveDefinitionIntoOrOutOfClass(
         }
     }
 
-    if (definition.parent?.isClassOrStruct()) {
+    if (definition.parent?.isClassType()) {
         const position = await getNewPosition(classDoc, definition);
 
         const definitionText = await definition.getDefinitionForTargetPosition(classDoc, position, declaration, true);
@@ -156,15 +156,15 @@ export async function moveDefinitionIntoOrOutOfClass(
         workspaceEdit.insert(classDoc.uri, position, formattedDefinition);
         const newDeclaration = definition.newFunctionDeclaration();
         const replaceRange = cfg.alwaysMoveComments(definition.uri)
-                ? definition.rangeWithLeadingComment()
-                : definition.fullRange();
+                ? definition.rangeWithComments
+                : definition.fullRange;
         workspaceEdit.replace(definition.uri, replaceRange, newDeclaration);
         return vscode.workspace.applyEdit(workspaceEdit);
     } else if (declaration) {
         const combinedDefinition = declaration.combineDefinition(definition);
 
         const workspaceEdit = new vscode.WorkspaceEdit();
-        workspaceEdit.replace(declaration.uri, declaration.fullRange(), combinedDefinition);
+        workspaceEdit.replace(declaration.uri, declaration.fullRange, combinedDefinition);
         const deletionRange = getDeletionRange(definition);
         workspaceEdit.delete(definition.uri, deletionRange);
         return vscode.workspace.applyEdit(workspaceEdit);
@@ -208,11 +208,13 @@ async function getNewPosition(targetDoc: SourceDocument, declaration?: SourceSym
 }
 
 function getDeletionRange(definition: CSymbol): vscode.Range {
-    let deletionRange = definition.rangeWithLeadingComment();
-    if (definition.document.lineAt(deletionRange.start.line - 1).isEmptyOrWhitespace) {
+    let deletionRange = definition.rangeWithComments;
+    if (deletionRange.start.line > 0
+            && definition.document.lineAt(deletionRange.start.line - 1).isEmptyOrWhitespace) {
         deletionRange = deletionRange.union(definition.document.lineAt(deletionRange.start.line - 1).range);
     }
-    if (definition.document.lineAt(deletionRange.end.line + 1).isEmptyOrWhitespace) {
+    if (deletionRange.end.line < definition.document.lineCount - 1
+            && definition.document.lineAt(deletionRange.end.line + 1).isEmptyOrWhitespace) {
         deletionRange = deletionRange.union(definition.document.lineAt(deletionRange.end.line + 1).range);
     }
     return deletionRange;